Refactoring DDSA glue code for easier debugging
